+++ apps/backoffice/src/shared/components/brand/BrandForm.tsx
@@ -4,47 +4,13 @@ import tw from 'tailwind-styled-components';
 
 import { LicenseFileField } from './LicenseFileField';
 
-import { FormField, Input, Select } from '@/shared/components';
+import { FormField, Input, Select, FieldWrapper } from '@/shared/components';
 import {
   registerBrandInputSchema,
   BusinessType,
   type RegisterBrandInput,
   type Brand,
 } from '@/types/brand.type';
-
-// 필드별 조회/수정 전환 컴포넌트
-interface FieldWrapperProps {
-  label: string;
-  value?: string | null;
-  isEditMode: boolean;
-  children: React.ReactNode;
-  required?: boolean;
-  error?: string;
-}
-
-function FieldWrapper({
-  label,
-  value,
-  isEditMode,
-  children,
-  required,
-  error,
-}: FieldWrapperProps) {
-  if (isEditMode) {
-    return (
-      <FormField label={label} error={error} required={required}>
-        {children}
-      </FormField>
-    );
-  }
-
-  return (
-    <InfoItem>
-      <InfoLabel>{label}</InfoLabel>
-      <InfoValue>{value || '-'}</InfoValue>
-    </InfoItem>
-  );
-}
 
 interface BrandFormProps {
   data?: Brand;
@@ -437,19 +403,4 @@ const EditButton = tw.button`
   hover:bg-blue-700
   transition-colors
   font-medium
-`;
-
-const InfoItem = tw.div`
-  space-y-1
-`;
-
-const InfoLabel = tw.dt`
-  text-sm
-  font-medium
-  text-gray-500
-`;
-
-const InfoValue = tw.dd`
-  text-sm
-  text-gray-900
 `;
