+++ apps/backoffice/src/shared/components/form/index.ts
@@ -1,3 +1,4 @@
 export * from './FormField';
 export * from './Input';
 export * from './Select';
+export * from './FieldWrapper';
